FIX tests

+++ src/controllers/helpers.coffee
@@ -4,13 +4,17 @@ module.exports =
   # rmdir
   # ======
   # Remove dir and files recursevely
-  rmdir: (path) ->
-    files = fs.readdirSync path
-    if files.length > 0
-      for file in files
-        filepath = path + "/" + file
-        if fs.statSync(filepath).isFile()
-          fs.unlinkSync filepath
-        else
-          rmdir filepath
-    fs.rmdirSync path
+  rmdir: (path, cb) ->
+    fs.readdir path, (err, files) ->
+      # Callback with error
+      return cb err if err
+      if files.length > 0
+        for file in files
+          filepath = path + "/" + file
+          if fs.statSync(filepath).isFile()
+            fs.unlinkSync filepath
+          else
+            rmdir filepath
+      fs.rmdirSync path
+      # callback
+      cb()
